BLOC_04/requirements.txt
# Dépendances pour l'exécution des DAGs et des tests

# ============================================================================
# AIRFLOW & ML
# ============================================================================
apache-airflow==2.8.0
pandas==2.2.2
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
mlflow==2.9.2

# ============================================================================
# AWS
# ============================================================================
boto3==1.34.10
moto==4.2.11  # Mock AWS services pour les tests

# ============================================================================
# API
# ============================================================================
requests==2.31.0
responses==0.24.1  # Mock HTTP pour les tests
requests-mock==1.11.0

# ============================================================================
# TESTS
# ============================================================================
pytest==8.3.2
pytest-cov==4.1.0
pytest-mock==3.14.0
pytest-html==4.1.1
pytest-xdist==3.5.0

# ============================================================================
# QUALITÉ DE CODE
# ============================================================================
flake8==6.1.0
pylint==3.0.3
black==23.12.1
isort==5.13.2

# ============================================================================
# UTILITAIRES TESTS
# ============================================================================
faker==20.1.0
factory-boy==3.3.0
